modularize views
