+++ frontend/src/modules/Datasets/views/DatasetList.js
@@ -19,10 +19,11 @@ import {
   PlusIcon,
   SearchInput,
   useSettings
-} from '../../../../design';
-import { SET_ERROR, useDispatch } from '../../../../globalErrors';
-import { listDatasets, useClient } from '../../../../services';
-import DatasetListItem from '../components/DatasetListItem';
+} from '../../../design';
+import { SET_ERROR, useDispatch } from '../../../globalErrors';
+import { useClient } from '../../../services';
+import { listDatasets } from '../services';
+import { DatasetListItem } from '../components';
 
 function DatasetsPageHeader() {
   return (
